refactor(core): moving the shared types to a crate
This reduces the server container dependencies significantly.
