+++ core/Cargo.toml
@@ -39,6 +39,8 @@ tracing-opentelemetry = { workspace = true }
 tracing-subscriber = { workspace = true }
 urlencoding = { workspace = true }
 
+ethos-types = { workspace = true }
+
 http = { version = "0.2.9" }
 
 aws-sdk-eks = { version = "1", features = ["behavior-version-latest"] }
