+++ core/src/types/aws.rs
@@ -1,10 +1,1 @@
-use chrono::{DateTime, Utc};
-use serde::{Deserialize, Serialize};
-
-#[derive(Debug, Deserialize, Serialize)]
-pub struct AWSStaticCredentials {
-    pub access_key_id: String,
-    pub secret_access_key: String,
-    pub session_token: Option<String>,
-    pub expiration: Option<DateTime<Utc>>,
-}
+pub use ethos_types::aws::AWSStaticCredentials;
