+++ friendshipper/server/Cargo.toml
@@ -5,14 +5,14 @@ edition = "2021"
 
 [dependencies]
 anyhow = "1.0.72"
-aws-config = "1.5.8"
+aws-config = { version = "1.5.8", features = ["behavior-version-latest"] }
 aws-sdk-sts = "1.46.0"
 aws-smithy-types-convert = { version = "0.60.8", features = ["convert-chrono"] }
 axum = { version = "0.7.5", features = ["macros"] }
 chrono = { version = "0.4.26", features = ["serde"] }
 clap = { version = "4.5.20", features = ["derive"] }
 config = "0.13.3"
-ethos-core = { path = "../../core" }
+ethos-types = { path = "../../types" }
 jwt-authorizer = "0.15.0"
 log = "0.4.20"
 opentelemetry = "0.21.0"
