+++ friendshipper/server/src/main.rs
@@ -1,6 +1,6 @@
 use axum::Router;
 use clap::Parser;
-use ethos_core::types::config::{FriendshipperConfig, OktaConfig};
+use ethos_types::config::{FriendshipperConfig, OktaConfig};
 use friendshipper_server::{ServerConfig, APP_NAME, VERSION};
 #[allow(unused_imports)]
 use jwt_authorizer::{Authorizer, IntoLayer, JwtAuthorizer, Refresh, RefreshStrategy};
