+++ types/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "ethos-types"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+anyhow = { workspace = true }
+axum = { workspace = true }
+chrono = { workspace = true }
+serde = { workspace = true }
